# [stable] Remove millions of duplicates from RuleReferencesRule