@dLobatog 2020-05-04
Debug logs for QE

```
2020-05-04T10:37:02.476Z 123 TID-golptandb INFO: GitLab reliable fetch activated!
D, [2020-05-04T10:37:02.580334 #123] DEBUG -- :    (0.2ms)  SELECT pg_try_advisory_lock(4483237340039375190)
D, [2020-05-04T10:37:02.580429 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
D, [2020-05-04T10:37:02.590381 #123] DEBUG -- :    (0.6ms)  SELECT "schema_migrations"."version" FROM "schema_migrations" ORDER BY "schema_migrations"."version" ASC
D, [2020-05-04T10:37:02.590496 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
I, [2020-05-04T10:37:02.590788 #123]  INFO -- : Migrating to UniquenessRuleReferencesRuleIndex (20200428205535)
D, [2020-05-04T10:37:02.594261 #123] DEBUG -- :    (0.4ms)  BEGIN
D, [2020-05-04T10:37:02.594385 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
== 20200428205535 UniquenessRuleReferencesRuleIndex: migrating ================
-- create_table(:new_rule_references_rules, {:id=>false})
D, [2020-05-04T10:37:02.599075 #123] DEBUG -- :    (1.4ms)  CREATE TABLE "new_rule_references_rules" ("rule_id" uuid NOT NULL, "rule_reference_id" uuid NOT NULL)
D, [2020-05-04T10:37:02.599232 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:8
D, [2020-05-04T10:37:02.605127 #123] DEBUG -- :    (4.3ms)  CREATE  INDEX  "index_new_rule_references_rules_on_rule_id" ON "new_rule_references_rules"  ("rule_id")
D, [2020-05-04T10:37:02.605219 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:8
D, [2020-05-04T10:37:02.609459 #123] DEBUG -- :    (2.9ms)  CREATE  INDEX  "index_new_rule_references_rules_on_rule_reference_id" ON "new_rule_references_rules"  ("rule_reference_id")
D, [2020-05-04T10:37:02.609557 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:8
   -> 0.0123s
D, [2020-05-04T10:37:02.694603 #123] DEBUG -- :    (82.6ms)  SELECT COUNT(*) FROM "rule_references_rules" WHERE "rule_references_rules"."id" IN (SELECT MIN(id) FROM "rule_references_rules" GROUP BY rule_id, rule_reference_id HAVING (count(id) > 1))
D, [2020-05-04T10:37:02.694722 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:15
D, [2020-05-04T10:37:02.776180 #123] DEBUG -- :    (77.3ms)  SELECT  "rule_references_rules"."id" FROM "rule_references_rules" WHERE "rule_references_rules"."id" IN (SELECT MIN(id) FROM "rule_references_rules" GROUP BY rule_id, rule_reference_id HAVING (count(id) > 1)) ORDER BY "rule_references_rules"."id" ASC LIMIT $1  [["LIMIT", 10000]]
D, [2020-05-04T10:37:02.776315 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:16
-- drop_table(:rule_references_rules)
D, [2020-05-04T10:37:02.779007 #123] DEBUG -- :    (2.3ms)  DROP TABLE "rule_references_rules"
D, [2020-05-04T10:37:02.779136 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:23
   -> 0.0026s
-- rename_table(:new_rule_references_rules, :rule_references_rules)
D, [2020-05-04T10:37:02.779661 #123] DEBUG -- :    (0.3ms)  ALTER TABLE "new_rule_references_rules" RENAME TO "rule_references_rules"
D, [2020-05-04T10:37:02.779790 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:24
D, [2020-05-04T10:37:02.789041 #123] DEBUG -- :    (0.4ms)  ALTER INDEX "index_new_rule_references_rules_on_rule_id" RENAME TO "index_rule_references_rules_on_rule_id"
D, [2020-05-04T10:37:02.789217 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:24
D, [2020-05-04T10:37:02.790360 #123] DEBUG -- :    (1.0ms)  ALTER INDEX "index_new_rule_references_rules_on_rule_reference_id" RENAME TO "index_rule_references_rules_on_rule_reference_id"
D, [2020-05-04T10:37:02.790545 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:24
   -> 0.0114s
-- add_index(:rule_references_rules, [:rule_id, :rule_reference_id], {:unique=>true})
D, [2020-05-04T10:37:02.797508 #123] DEBUG -- :    (4.0ms)  CREATE UNIQUE INDEX  "index_rule_references_rules_on_rule_id_and_rule_reference_id" ON "rule_references_rules"  ("rule_id", "rule_reference_id")
D, [2020-05-04T10:37:02.797727 #123] DEBUG -- :   ↳ db/migrate/20200428205535_uniqueness_rule_references_rule_index.rb:25
   -> 0.0071s
== 20200428205535 UniquenessRuleReferencesRuleIndex: migrated (0.2033s) =======

D, [2020-05-04T10:37:02.801600 #123] DEBUG -- :   ActiveRecord::SchemaMigration Create (0.5ms)  INSERT INTO "schema_migrations" ("version") VALUES ($1) RETURNING "version"  [["version", "20200428205535"]]
D, [2020-05-04T10:37:02.801833 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
D, [2020-05-04T10:37:02.806023 #123] DEBUG -- :    (3.9ms)  COMMIT
D, [2020-05-04T10:37:02.806419 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
D, [2020-05-04T10:37:02.817112 #123] DEBUG -- :   ActiveRecord::InternalMetadata Load (0.5ms)  SELECT  "ar_internal_metadata".* FROM "ar_internal_metadata" WHERE "ar_internal_metadata"."key" = $1 LIMIT $2  [["key", "environment"], ["LIMIT", 1]]
D, [2020-05-04T10:37:02.817534 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
D, [2020-05-04T10:37:02.825306 #123] DEBUG -- :    (0.4ms)  BEGIN
D, [2020-05-04T10:37:02.825507 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
D, [2020-05-04T10:37:02.826576 #123] DEBUG -- :    (0.3ms)  COMMIT
D, [2020-05-04T10:37:02.826699 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
D, [2020-05-04T10:37:02.827132 #123] DEBUG -- :    (0.3ms)  SELECT pg_advisory_unlock(4483237340039375190)
D, [2020-05-04T10:37:02.827276 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
D, [2020-05-04T10:37:02.829527 #123] DEBUG -- :    (0.5ms)  SELECT "schema_migrations"."version" FROM "schema_migrations" ORDER BY "schema_migrations"."version" ASC
D, [2020-05-04T10:37:02.829675 #123] DEBUG -- :   ↳ /usr/local/bundle/bin/rake:23
```